@@ -1,3 +1,5 @@
+import type {SidebarsConfig} from '@docusaurus/plugin-content-docs';
+
 /**
  * Creating a sidebar enables you to:
  - create an ordered group of docs
@@ -8,27 +10,22 @@
 
  Create as many sidebars as you want.
  */
-
-// @ts-check
-
-/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
-const sidebars = {
+const sidebars: SidebarsConfig = {
   // By default, Docusaurus generates a sidebar from the docs folder structure
-  // tutorialSidebar: [{type: 'autogenerated', dirName: '.'}],
+  tutorialSidebar: [{type: 'autogenerated', dirName: '.'}],
 
   // But you can create a sidebar manually
+  /*
   tutorialSidebar: [
+    'intro',
+    'hello',
     {
-      type: 'autogenerated',
-      dirName: '.',
+      type: 'category',
+      label: 'Tutorial',
+      items: ['tutorial-basics/create-a-document'],
     },
-    // {
-    //   type: 'link',
-    //   label: 'Javadoc ',
-    //   // As docusaurus translate http prexixed url as external
-    //   href: 'http://localhost:3000/javadoc/index.html'
-    // }
   ],
+   */
 };
 
-module.exports = sidebars;
+export default sidebars;
